Add showHint option to override `first` value
Passing first=true to `update` makes sense when the user is explicitly
requesting completions, so that if there is only one item in the list it
can complete their text.

It makes less sense if the editor is configured to show hints as often
as possible. `options.firstOverride` provides a way for such an editor
to prevent auto-completion when all we want is to show the list as the
user types, even if there is only one item.